Extract cifar10 TF model from model/Model.py
